Match & link CFluidPlane
