+++ config/GM8E01_00/splits.txt
@@ -881,13 +881,18 @@ MetroidPrime/Cameras/CPathCamera.cpp:
 	.sbss       start:0x805A8F60 end:0x805A8F70
 	.sdata2     start:0x805AB5C8 end:0x805AB5E8
 
-MetroidPrime/CFluidPlaneManager.cpp:
-	.text       start:0x8012E90C end:0x8012F774
-	.rodata     start:0x803D0038 end:0x803D0048
+MetroidPrime/CFluidPlane.cpp:
+	.text       start:0x8012E90C end:0x8012F098
 	.data       start:0x803E0CE0 end:0x803E0D00
+	.sdata      start:0x805A76C8 end:0x805A76D8
+	.sdata2     start:0x805AB5E8 end:0x805AB608
+
+MetroidPrime/CFluidPlaneManager.cpp:
+	.text       start:0x8012F098 end:0x8012F774
+	.rodata     start:0x803D0038 end:0x803D0048
 	.bss        start:0x8056F8B4 end:0x805719CC align:4 common
-	.sdata      start:0x805A76C8 end:0x805A76E0
-	.sdata2     start:0x805AB5E8 end:0x805AB650
+	.sdata      start:0x805A76D8 end:0x805A76E0
+	.sdata2     start:0x805AB608 end:0x805AB650
 
 MetroidPrime/ScriptObjects/CScriptGrapplePoint.cpp:
 	.text       start:0x8012F774 end:0x8012FAD8
